requirements.txt
fastapi==0.128.0
uvicorn==0.42.0
sounddevice
pydantic==2.12.5
pydantic-settings==2.12.0

openai==2.28.0
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.0
langchain==1.1.2
langchain-text-splitters==1.0.0
langchain-core==1.2.19
langchain-community==0.4.1
langchain-openai==1.1.11
langgraph==1.0.4
langsmith==0.4.56
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
dateparser==1.2.2
DateTime==6.0
httpx==0.28.1
requests==2.32.5
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
python-dotenv==1.2.1

deepgram-sdk==6.0.1
elevenlabs==2.39.0

redis==7.1.0
SQLAlchemy==2.0.44
websockets
orjson==3.11.5
numpy
